bundled first example into own subdirectory with makefile for easier compilation and ability to add more examples later
