+++ .gitignore
@@ -1,3 +1,4 @@
 build/*
 .idea/
 .project
+examples/**/build
