+++ examples/cli_with_2_services/svc1/httpclient.go
