+++ examples/cli_with_2_services/svc1/service.go
