+++ examples/cli_with_2_services/svc2/cmd/main.go
@@ -10,16 +10,15 @@ import (
 	"github.com/opentracing/opentracing-go"
 
 	zipkin "github.com/openzipkin/zipkin-go-opentracing"
-	"github.com/openzipkin/zipkin-go-opentracing/examples/svc1"
-	"github.com/openzipkin/zipkin-go-opentracing/examples/svc2"
+	"github.com/openzipkin/zipkin-go-opentracing/examples/cli_with_2_services/svc2"
 )
 
 const (
 	// Our service name.
-	serviceName = "svc1"
+	serviceName = "svc2"
 
 	// Host + port of our service.
-	hostPort = "127.0.0.1:61001"
+	hostPort = "127.0.0.1:61002"
 
 	// Endpoint to send Zipkin spans to.
 	zipkinHTTPEndpoint = "http://localhost:9411/api/v1/spans"
@@ -27,14 +26,11 @@ const (
 	// Debug mode.
 	debug = false
 
-	// Base endpoint of our SVC1 service.
-	svc2Endpoint = "http://localhost:61002"
-
 	// same span can be set to true for RPC style spans (Zipkin V1) vs Node style (OpenTracing)
 	sameSpan = true
 )
 
-//svc1
+//svc2
 func main() {
 	// create collector.
 	collector, err := zipkin.NewHTTPCollector(zipkinHTTPEndpoint)
@@ -58,14 +54,11 @@ func main() {
 	// explicitely set our tracer to be the default tracer.
 	opentracing.InitGlobalTracer(tracer)
 
-	// create the client to svc2
-	svc2Client := svc2.NewHTTPClient(tracer, svc2Endpoint)
-
 	// create the service implementation
-	service := svc1.NewService(svc2Client)
+	service := svc2.NewService()
 
 	// create the HTTP Server Handler for the service
-	handler := svc1.NewHTTPHandler(tracer, service)
+	handler := svc2.NewHTTPHandler(tracer, service)
 
 	// start the service
 	fmt.Printf("Starting %s on %s\n", serviceName, hostPort)
