+++ examples/cli_with_2_services/svc2/httpclient.go
