+++ examples/cli_with_2_services/svc2/httpserver.go
@@ -1,6 +1,6 @@
 // +build go1.7
 
-package svc1
+package svc2
 
 import (
 	"fmt"
@@ -8,24 +8,12 @@ import (
 	"strconv"
 
 	opentracing "github.com/opentracing/opentracing-go"
+
 	"github.com/openzipkin/zipkin-go-opentracing/examples/middleware"
 )
 
 type httpService struct {
 	service Service
-}
-
-// concatHandler is our HTTP HandlerFunc for a Concat request.
-func (s *httpService) concatHandler(w http.ResponseWriter, req *http.Request) {
-	// parse query parameters
-	v := req.URL.Query()
-	result, err := s.service.Concat(req.Context(), v.Get("a"), v.Get("b"))
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-	// return the result
-	w.Write([]byte(result))
 }
 
 // sumHandler is our HTTP Handlerfunc for a Sum request.
@@ -52,20 +40,13 @@ func (s *httpService) sumHandler(w http.ResponseWriter, req *http.Request) {
 	w.Write([]byte(fmt.Sprintf("%d", result)))
 }
 
-// NewHTTPHandler returns a new HTTP handler our svc1.
+// NewHTTPHandler returns a new HTTP handler our svc2.
 func NewHTTPHandler(tracer opentracing.Tracer, service Service) http.Handler {
 	// Create our HTTP Service.
 	svc := &httpService{service: service}
 
 	// Create the mux.
 	mux := http.NewServeMux()
-
-	// Create the Concat handler.
-	var concatHandler http.Handler
-	concatHandler = http.HandlerFunc(svc.concatHandler)
-
-	// Wrap the Concat handler with our tracing middleware.
-	concatHandler = middleware.FromHTTPRequest(tracer, "Concat")(concatHandler)
 
 	// Create the Sum handler.
 	var sumHandler http.Handler
@@ -75,7 +56,6 @@ func NewHTTPHandler(tracer opentracing.Tracer, service Service) http.Handler {
 	sumHandler = middleware.FromHTTPRequest(tracer, "Sum")(sumHandler)
 
 	// Wire up the mux.
-	mux.Handle("/concat/", concatHandler)
 	mux.Handle("/sum/", sumHandler)
 
 	// Return the mux.
