+++ examples/cli_with_2_services/svc2/service.go
